Conflicts:
	kura/emulator/org.eclipse.kura.emulator/src/main/java/org/eclipse/kura/emulator/Emulator.java
	kura/examples/org.eclipse.kura.example.serial.publisher/src/main/java/org/eclipse/kura/example/serial/publisher/ExampleSerialPublisher.java
	kura/org.eclipse.kura.ble.provider/src/main/java/org/eclipse/kura/internal/ble/BluetoothLeServiceImpl.java
	kura/org.eclipse.kura.camel.cloud.factory/src/org/eclipse/kura/camel/cloud/factory/internal/CamelCloudServiceFactory.java
	kura/org.eclipse.kura.camel.cloud.factory/src/org/eclipse/kura/camel/cloud/factory/internal/CamelFactory.java
	kura/org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider/src/main/java/org/eclipse/kura/internal/cloudconnection/eclipseiot/mqtt/cloud/factory/DefaultCloudConnectionFactory.java
	kura/org.eclipse.kura.cloudconnection.raw.mqtt.provider/src/main/java/org/eclipse/kura/cloudconnection/raw/mqtt/factory/RawMqttCloudConnectionFactory.java
	kura/org.eclipse.kura.core.configuration/src/main/java/org/eclipse/kura/core/configuration/ConfigurationServiceImpl.java
	kura/org.eclipse.kura.core.configuration/src/main/java/org/eclipse/kura/core/configuration/metatype/Tocd.java
	kura/org.eclipse.kura.core/src/main/java/org/eclipse/kura/core/data/DataMessage.java
	kura/org.eclipse.kura.driver.opcua.provider/src/main/java/org/eclipse/kura/internal/driver/opcua/OpcUaDriver.java
	kura/org.eclipse.kura.driver.opcua.provider/src/main/java/org/eclipse/kura/internal/driver/opcua/SubscriptionManager.java
	kura/org.eclipse.kura.driver.s7plc.provider/src/main/java/org/eclipse/kura/internal/driver/s7plc/S7PlcDomain.java
	kura/org.eclipse.kura.driver.s7plc.provider/src/main/java/org/eclipse/kura/internal/driver/s7plc/task/S7PlcToplevelBlockTask.java
	kura/org.eclipse.kura.json.marshaller.unmarshaller.provider/src/main/java/org/eclipse/kura/internal/json/marshaller/unmarshaller/message/CloudPayloadJsonDecoder.java
	kura/org.eclipse.kura.linux.clock/OSGI-INF/metatype/org.eclipse.kura.clock.ClockService.xml
	kura/org.eclipse.kura.protocol.modbus/src/main/java/org/eclipse/kura/protocol/modbus/ModbusProtocolDevice.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/EntryClassUi.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/login/LoginUi.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/status/StatusPanelUi.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/wires/WireComponentDescriptors.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/client/ui/wires/WiresDialogs.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/server/servlet/DeviceSnapshotsServlet.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/server/servlet/WiresBlinkServlet.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtCloudEntry.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/model/GwtConfigComponent.java
	kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/service/GwtComponentService.java
	kura/org.eclipse.kura.wire.component.provider/src/main/java/org/eclipse/kura/internal/wire/asset/RecordFillers.java
	kura/org.eclipse.kura.wire.script.filter.provider/src/main/java/org/eclipse/kura/wire/script/filter/provider/ScriptFilter.java
	kura/pom.xml
	kura/test/org.eclipse.kura.test/src/org/eclipse/kura/test/RemoteTargetTest.java
	kura/test/org.eclipse.kura.test/src/org/eclipse/kura/test/TestExtender.java
	target-platform/com.gwt.user/pom.xml
	target-platform/org.usb4java/pom.xml
	target-platform/usb4java-javax/pom.xml

+++ kura/org.eclipse.kura.core/src/main/java/org/eclipse/kura/core/internal/linux/executor/ExecutorUtil.java
@@ -298,7 +298,7 @@ public class ExecutorUtil {
     }
 
     private static String[] buildKillCommand(Pid pid, Signal signal) {
-        Integer pidNumber = (Integer) pid.getPid();
+        Integer pidNumber = pid.getPid();
         if (logger.isInfoEnabled()) {
             logger.info("Attempting to send {} to process with pid {}", ((LinuxSignal) signal).name(), pidNumber);
         }
